+++ explorer/admin/ExplorerBaker.tsx
@@ -13,7 +13,7 @@ import { ExplorerProps } from "explorer/client/Explorer"
 import { FunctionalRouter } from "adminSite/server/utils/FunctionalRouter"
 import { getGrapherById } from "db/model/Chart"
 import { Router } from "express"
-import { GIT_CMS_DIR } from "gitCms/constants"
+import { GIT_CMS_DIR } from "gitCms/GitCmsConstants"
 import { getBlockContent } from "db/wpdb"
 import { ExplorerPage } from "./ExplorerPage"
 import moment from "moment"
