+++ explorer/admin/ExplorerCreatePage.tsx
@@ -11,7 +11,7 @@ import { action, observable, computed } from "mobx"
 import { GrapherInterface } from "grapher/core/GrapherInterface"
 import { ExplorerProgram } from "explorer/client/ExplorerProgram"
 import { DefaultExplorerProgram } from "explorer/client/DefaultExplorerProgram"
-import { readRemoteFile, writeRemoteFile } from "gitCms/client"
+import { readRemoteFile, writeRemoteFile } from "gitCms/GitCmsClient"
 import { Prompt } from "react-router-dom"
 import { Link } from "adminSite/client/Link"
 import Handsontable from "handsontable"
